pyproject.toml
[project]
name = "gijirog"
version = "0.1.0"
description = "Discord bot for meeting minutes"
readme = "README.md"
authors = [
    { name = "gafafrog", email = "code@gafafrog.com" }
]
requires-python = ">=3.12"
dependencies = [
    "discord-py>=2.7.1",
]

[project.scripts]
gijirog = "gijirog:main"

[build-system]
requires = ["uv_build>=0.11.7,<0.12.0"]
build-backend = "uv_build"
